pyproject.toml
[project]
name = "uv-docker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=1.26.4",
    "ipykernel>=6.29.5",
    "hydra-core>=1.3.2",
    "polars>=1.7.1",
    "scikit-learn>=1.5.2",
    "lightgbm>=4.5.0",
    "pandas>=2.2.2",
    "pyarrow>=17.0.0",
    "seaborn>=0.13.2",
    "matplotlib>=3.9.2",
    # torchを使う際は、以下のコメントアウトを解除
    "torch==2.4.0+cu121; sys_platform == 'linux' and platform_machine == 'x86_64'",
    "torch==2.4.0; sys_platform == 'darwin' or (sys_platform == 'linux' and platform_machine == 'aarch64')",
    "torchvision==0.19.0+cu121; sys_platform == 'linux'",
    "torchvision==0.19.0; sys_platform == 'darwin' or (sys_platform == 'linux' and platform_machine == 'aarch64')",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.3",
    "ruff>=0.6.5",
    "mypy>=1.11.2",
]
find-links = [
    "https://download.pytorch.org/whl/cu121/torch",
    "https://download.pytorch.org/whl/cu121/torchvision",
]

[tool.ruff]
exclude = [".git", ".mypy_cache", ".ruff_cache", ".venv", "third_party"]
line-length = 120
indent-width = 4
target-version = "py312"

[tool.ruff.format]
quote-style = "double"
line-ending = "auto"

[tool.ruff.lint]
fixable = ["ALL"]
unfixable = []
select = [
    "A",  # flake8-builtin
    "B",  # flake8-bugbear
    "E",  # pycodestyle error
    "F",  # Pyflakes
    "I",  # isort
    "N",  # pep8-naming
    "W",  # pycodestyle warning
    "PL",  # Pylint
    "UP",  # pyupgrade
]
ignore = [
    "F401",  # Unused import
    "PLR0913",  # Too many arguments
    "PLR2004",  # Magic value used in comparison, consider replacing with a constant
    "PLW2901",  # `for` loop variable overwritten by assignment target
    "N803",  # Argument name should be lowercase
    "N806",  # Variable name should be lowercase
    "N812",  # Lowercase imported as non-lowercase
]
